wt/pyproject.toml
[project]
name = "wt"
version = "0.0.0"
description = "Ecoscope Workflows CLI - Tool for managing ecoscope repositories"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-3-Clause" }

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pygithub>=2.3.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
wt = "wt.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wt"]
only-packages = true

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-arm64", "osx-64", "linux-64", "win-64"]

[tool.pixi.pypi-dependencies]
wt = { path = ".", editable = true }

[tool.pixi.dependencies]
python = ">=3.10"
hatchling = "*"
